refactor: organize files under /lib
