+++ lib/rules/no_descendant_combinator.ts
@@ -1,5 +1,5 @@
 import stylelint from "stylelint";
-import { walkSelectors } from "./helpers/walk_selectors";
+import { walkSelectors } from "../helpers/walk_selectors";
 
 const { utils, createPlugin } = stylelint;
